product-service/requirements.txt
Flask==3.0.0
flask-cors==5.0.0
requests==2.31.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
